Fixed the Goddamn puzzle again. Oh and also shooting
shooting is now a bool that returns true if it can fire and false otherwise. this allows us to handle animations in the handler and logic in the shoot. Also the puzzle is now a little different. it remembers where you last had your sine wave, however failing regardless sets it to zero.