Stupid ground check
I am losing my mind trying to figure out movement so here is what I have. It is garbage and is doing what the old one was doing. a funny CharacterController.Move, This is a really bad way of doing it and needs to be tested in a scene to see if it jitters. Literally don't understand collisions anymore.