Lots of changes to UI, Dialogue UI, Inventory UI, not really health UI and some quality of life publics to privates
Big changes to UI, and Inventory scripts. they now are more private and allow for less initialization. WARNING: the UI Manager is vital per scene, note that Inventory is not broken, it has been hacked into the Input manager, (hit tab) it is not a current or for the future feature. lastly, note that UI itself, needs health icon the way it is set currently, everything else is private with getters for easy use and for scripts that you don't like initializing every time you use them.