Changed the script execution order, and changed the player
Now the Scripts execute correctly no more null reference.
Next, a change in the player controller that I found to work thanks to Cameron. We now use rigidbodies like planned, it was being weird because we set it to is kinematic and because mixamo rigs are generally not great. All in all fixed and changed