major area tweaks and prefabs
Slight tweaking of major areas to have empty game objects for areas of interest and separation from one another. converted all major areas in the main scene to prefabs, they can be worked on instead of the main scene for testing purpose and for making us less dependent on one another.